Extract each macro into individual unit test suite
