+++ test/unit/macros/common/from-now.test.js
