+++ test/unit/macros/common/progress.test.js
